Comment unnecessary config
I need to comments that parts of config to dev server works
correct(without that comments I get an error because these assets do
not exist now, but I do not want to delete it from cofnig, because
these configuration entries will be useful in the future when I create
subpages/add imgs to page). The same concerns prod config.